+++ packages/presets/src/theme/recipes/menu.ts
@@ -4,7 +4,6 @@ import { defineSlotRecipe } from '@pandacss/dev'
 const itemStyle = {
   alignItems: 'center',
   borderRadius: 'l1',
-  color: 'fg.emphasized',
   cursor: 'pointer',
   display: 'flex',
   fontWeight: 'medium',
@@ -13,12 +12,16 @@ const itemStyle = {
   transitionProperty: 'background, color',
   transitionTimingFunction: 'default',
   _hover: {
-    color: 'fg.default',
     background: 'bg.subtle',
+    '& :where(svg)': {
+      color: 'fg.default',
+    },
   },
   _highlighted: {
-    color: 'fg.default',
     background: 'bg.subtle',
+  },
+  '& :where(svg)': {
+    color: 'fg.muted',
   },
 } as const
 
@@ -39,6 +42,7 @@ export const menu = defineSlotRecipe({
       display: 'flex',
       flexDirection: 'column',
       outline: 'none',
+      width: 'calc(100% + 2rem)',
       _hidden: {
         display: 'none',
       },
@@ -74,22 +78,33 @@ export const menu = defineSlotRecipe({
         content: {
           py: '1',
           gap: '1',
-          minW: '24',
         },
         item: {
           h: '8',
           px: '1.5',
           mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
         optionItem: {
           h: '8',
           px: '1.5',
           mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
         triggerItem: {
           h: '8',
           px: '1.5',
-          mx: '1.5',
+          mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
       },
       sm: {
@@ -104,22 +119,33 @@ export const menu = defineSlotRecipe({
         content: {
           py: '1',
           gap: '1',
-          minW: '32',
         },
         item: {
           h: '9',
           px: '2',
           mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
         optionItem: {
           h: '9',
           px: '2',
           mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
         triggerItem: {
           h: '9',
           px: '2',
           mx: '1.5',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
       },
       md: {
@@ -134,22 +160,74 @@ export const menu = defineSlotRecipe({
         content: {
           py: '1',
           gap: '1',
-          minW: '44',
         },
         item: {
           h: '10',
           px: '2.5',
           mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
         optionItem: {
           h: '10',
           px: '2.5',
           mx: '1',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
         },
         triggerItem: {
           h: '10',
           px: '2.5',
           mx: '1.5',
+          '& :where(svg)': {
+            width: '4',
+            height: '4',
+          },
+        },
+      },
+      lg: {
+        itemGroup: {
+          gap: '1',
+        },
+        itemGroupLabel: {
+          py: '2.5',
+          px: '2.5',
+          mx: '1',
+        },
+        content: {
+          py: '1',
+          gap: '1',
+        },
+        item: {
+          h: '11',
+          px: '2.5',
+          mx: '1',
+          '& :where(svg)': {
+            width: '5',
+            height: '5',
+          },
+        },
+        optionItem: {
+          h: '11',
+          px: '2.5',
+          mx: '1',
+          '& :where(svg)': {
+            width: '5',
+            height: '5',
+          },
+        },
+        triggerItem: {
+          h: '11',
+          px: '2.5',
+          mx: '1.5',
+          '& :where(svg)': {
+            width: '5',
+            height: '5',
+          },
         },
       },
     },