+++ website/src/content/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Added
+
+- Added `ghost` variant to the `Select` component.
+
 ## [0.12.0] - 2023-09-14
 
 ### Added